backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy>=1.26.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sshtunnel==0.4.0
yfinance>=0.2.0
